+++ frontend/src/pages/ExploreData/DefaultHelperBox.module.scss
@@ -1,0 +1,125 @@
+@use "../../styles/globals.scss";
+@import "../../styles/variables.module";
+
+.NoTopicContent {
+  margin: 1rem;
+  width: 100%;
+
+  padding: 1rem 3rem 0 3rem;
+
+  @media (min-width: $sm) {
+    padding: 2rem 10rem 0 10rem;
+  }
+}
+
+.NoTopicAlert {
+  justify-content: center !important;
+  align-items: center !important;
+  align-content: center !important;
+  justify-items: center !important;
+  margin: 0 0 20rem 0;
+}
+
+.NoTopicBox {
+  // margin: 0 3rem;
+  width: 70vw;
+}
+
+.BigHeadline {
+  display: none;
+
+  @media (min-width: $sm) {
+    display: block;
+  }
+
+  font-size: $text;
+  padding-right: 1rem;
+
+  @media (min-width: $md) {
+    font-size: $bigHeader;
+  }
+}
+
+.LittleHeadline {
+  font-size: $small;
+  padding-right: 1rem;
+
+  @media (min-width: $md) {
+    font-size: $smallestHeader;
+  }
+}
+
+.NoTopicHelperVideoBoxWithCaption {
+  margin-top: 1rem;
+}
+
+.NoTopicHelperVideoBox {
+  iframe {
+    height: 157px;
+    width: 250px;
+  }
+
+  @media (min-width: $lg) {
+    iframe {
+      height: 252px;
+      width: 400px;
+    }
+  }
+
+  @media (min-width: $xl) {
+    iframe {
+      height: 378px;
+      width: 600px;
+    }
+  }
+}
+
+.CategoryList {
+  ul span,
+  ul h2 {
+    display: flex;
+    font-size: $smallest !important;
+    line-height: $lhTight;
+
+    @media only screen and (min-width: $md) {
+      font-size: $text !important;
+      line-height: $lhLoose;
+    }
+  }
+}
+
+.ResourceVideoEmbed {
+  border-radius: 30px;
+}
+
+.SuggestedReportsList {
+  margin-block: 5px;
+
+  text-align: left;
+  font-size: $text;
+
+  list-style: none;
+
+  li::before {
+    content: "→ ";
+    color: $alt-green;
+  }
+
+  padding-left: 0px;
+
+  @media only screen and (min-width: $sm) {
+    padding-left: 40px;
+  }
+}
+
+SuggestedReportsListItem {
+  line-height: $lhNormal;
+}
+
+.SuggestedReportsListItem a {
+  text-decoration: none;
+
+  @media only screen and (min-width: $sm) {
+    line-height: $lhSuperLoose;
+  }
+}
